This is a work in progress commit only, NOT FOR TESTING!!!!!! Note that there is a separate release plan for Dream 1.7 so this experimental branch will be re-targetted to 2.0. the 1.7testing tag will be removed after this commit.

+++ common/MDI/MDIDecode.cpp
@@ -136,7 +136,7 @@ void CDecodeRSIMDI::InitInternal(CParameter& ReceiverParam)
 		iOutputBlockSize = NUM_FAC_BITS_PER_BLOCK;
 		//iOutputBlockSize2 = ReceiverParam.iNumSDCBitsPerSFrame;
 		iMaxOutputBlockSize2 = 1024;
-		int numstreams = ReceiverParam.Stream.Size();
+		size_t numstreams = size_t(ReceiverParam.Stream.Size());
 		//vecpvecOutputData.resize(numstreams);
 		for(size_t i=0; i<numstreams; i++)
 		{
